requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Limiter
python-dotenv
qrcode
pyotp
Werkzeug
